Playing around with Decimal layout

+++ src/jvmTest/kotlin/com/decimal128/decimal/BigDecimalToIEEE754.kt
@@ -12,11 +12,11 @@ private const val EMAX  =  6144
 private const val QTINY = EMIN - (P34 - 1) // -6176
 private const val QMAX  = EMAX - (P34 - 1) // 6111
 
-private const val INFINITY_SCALE = Integer.MAX_VALUE - 2
+private const val INFINITY_SCALE = Short.MAX_VALUE.toInt() - 2
 private val POS_INFINITY_SURROGATE = BigDecimal.ONE.scaleByPowerOfTen(INFINITY_SCALE)
 private val NEG_INFINITY_SURROGATE = POS_INFINITY_SURROGATE.negate()
-private const val QNAN_SCALE = Integer.MAX_VALUE - 1
-private const val SNAN_SCALE = Integer.MAX_VALUE
+private const val QNAN_SCALE = Short.MAX_VALUE.toInt() - 1
+private const val SNAN_SCALE = Short.MAX_VALUE.toInt()
 private val MAX_FINITE =
     BigDecimal.ONE.scaleByPowerOfTen(34).subtract(BigDecimal.ONE).scaleByPowerOfTen(6144-33)
 private val NEG_MAX_FINITE = MAX_FINITE.negate()
